feat: auctions
fix gen

fix auction
