+++ js/packages/api/src/index.ts
@@ -1,22 +1,26 @@
-import { DEFAULT_ENDPOINT } from './api';
+import { ENDPOINTS, MetaplexApi } from './api';
 import { inspect } from 'util';
 import { Connection } from '@solana/web3.js';
 import { loadAccounts } from '@oyster/common/dist/lib/contexts/meta/loadAccounts';
 import { getOwnedMeta } from '@oyster/common/dist/lib/contexts/meta/getOwnedMeta';
-import { loadUserTokenAccounts } from './loaders/loadUserTokenAccounts';
+import { loadUserTokenAccounts } from './utils/loadUserTokenAccounts';
 
 const ownerId = '7DAjuKLx8yHgcqfvTXaPwx5pUhDpWrF9XoTbBgT2fHrm';
 
 export const getData = async (): Promise<any> => {
-  const endpoint = DEFAULT_ENDPOINT.endpoint;
-  const connection = new Connection(endpoint, 'recent');
+  // const endpoint = ENDPOINTS[2].endpoint;
+  // const connection = new Connection(endpoint, 'recent');
 
-  const [state, userAccounts] = await Promise.all([
-    loadAccounts(connection, true),
-    loadUserTokenAccounts(connection, ownerId),
-  ]);
+  // const [state, userAccounts] = await Promise.all([
+  //   loadAccounts(connection, true),
+  //   loadUserTokenAccounts(connection, ownerId),
+  // ]);
 
-  return getOwnedMeta(userAccounts, state);
+  // return getOwnedMeta(userAccounts, state);
+  // return state;
+  const api = new MetaplexApi();
+  await api.initialize({ context: {} });
+  return (await api.getAuctions()).slice(0, 10);
 };
 
 const run = async () => {
