+++ js/packages/common/src/models/auction.ts
@@ -1,3 +1,11 @@
+import BN from 'bn.js';
+import {
+  AuctionData,
+  AuctionState,
+  BidderMetadata,
+  BidderPot,
+} from '../actions/auction';
+import { SafetyDepositBox, Vault } from '../actions/vault';
 import {
   Edition,
   MasterEditionV1,
@@ -8,7 +16,15 @@ import { ParsedAccount } from '../contexts/accounts/types';
 import { StringPublicKey } from '../utils/ids';
 import {
   AmountRange,
+  AuctionManager,
+  AuctionManagerStatus,
+  AuctionManagerV1,
+  AuctionManagerV2,
+  AuctionViewItem,
+  BidRedemptionTicket,
+  BidRedemptionTicketV2,
   ParticipationConfigV2,
+  SafetyDepositConfig,
   WinningConfigType,
 } from './metaplex';
 
@@ -21,4 +37,292 @@ export interface SafetyDepositDraft {
   winningConfigType: WinningConfigType;
   amountRanges: AmountRange[];
   participationConfig?: ParticipationConfigV2;
+}
+
+export enum AuctionViewState {
+  Live = '0',
+  Upcoming = '1',
+  Ended = '2',
+  BuyNow = '3',
+  Defective = '-1',
+}
+
+// Flattened surface item for easy display
+export interface AuctionView {
+  // items 1:1 with winning configs FOR NOW
+  // once tiered auctions come along, this becomes an array of arrays.
+  items: AuctionViewItem[][];
+  safetyDepositBoxes: ParsedAccount<SafetyDepositBox>[];
+  auction: ParsedAccount<AuctionData>;
+  auctionManager: AuctionManager;
+  participationItem?: AuctionViewItem;
+  state: AuctionViewState;
+  thumbnail: AuctionViewItem;
+  myBidderMetadata?: ParsedAccount<BidderMetadata>;
+  myBidderPot?: ParsedAccount<BidderPot>;
+  myBidRedemption?: ParsedAccount<BidRedemptionTicket>;
+  vault: ParsedAccount<Vault>;
+  totallyComplete: boolean;
+}
+
+export function buildListWhileNonZero<T>(hash: Record<string, T>, key: string) {
+  const list: T[] = [];
+  let ticket = hash[key + '-0'];
+  if (ticket) {
+    list.push(ticket);
+    let i = 1;
+    while (ticket) {
+      ticket = hash[key + '-' + i.toString()];
+      if (ticket) list.push(ticket);
+      i++;
+    }
+  }
+  return list;
+}
+
+export function processAccountsIntoAuctionView(
+  walletPubkey: StringPublicKey | null | undefined,
+  auction: ParsedAccount<AuctionData>,
+  auctionManagersByAuction: Record<
+    string,
+    ParsedAccount<AuctionManagerV1 | AuctionManagerV2>
+  >,
+  safetyDepositBoxesByVaultAndIndex: Record<
+    string,
+    ParsedAccount<SafetyDepositBox>
+  >,
+  metadataByMint: Record<string, ParsedAccount<Metadata>>,
+  bidderMetadataByAuctionAndBidder: Record<
+    string,
+    ParsedAccount<BidderMetadata>
+  >,
+  bidderPotsByAuctionAndBidder: Record<string, ParsedAccount<BidderPot>>,
+  bidRedemptionV2sByAuctionManagerAndWinningIndex: Record<
+    string,
+    ParsedAccount<BidRedemptionTicketV2>
+  >,
+  masterEditions: Record<
+    string,
+    ParsedAccount<MasterEditionV1 | MasterEditionV2>
+  >,
+  vaults: Record<string, ParsedAccount<Vault>>,
+  safetyDepositConfigsByAuctionManagerAndIndex: Record<
+    string,
+    ParsedAccount<SafetyDepositConfig>
+  >,
+  masterEditionsByPrintingMint: Record<string, ParsedAccount<MasterEditionV1>>,
+  masterEditionsByOneTimeAuthMint: Record<
+    string,
+    ParsedAccount<MasterEditionV1>
+  >,
+  metadataByMasterEdition: Record<string, ParsedAccount<Metadata>>,
+  cachedRedemptionKeysByWallet: Record<
+    string,
+    ParsedAccount<BidRedemptionTicket> | { pubkey: StringPublicKey; info: null }
+  >,
+  desiredState: AuctionViewState | undefined,
+  existingAuctionView?: AuctionView,
+): AuctionView | undefined {
+  let state: AuctionViewState;
+  if (auction.info.ended()) {
+    state = AuctionViewState.Ended;
+  } else if (auction.info.state === AuctionState.Started) {
+    state = AuctionViewState.Live;
+  } else if (auction.info.state === AuctionState.Created) {
+    state = AuctionViewState.Upcoming;
+  } else {
+    state = AuctionViewState.BuyNow;
+  }
+
+  const auctionManagerInstance = auctionManagersByAuction[auction.pubkey || ''];
+
+  // The defective auction view state really applies to auction managers, not auctions, so we ignore it here
+  if (
+    desiredState &&
+    desiredState !== AuctionViewState.Defective &&
+    desiredState !== state
+  )
+    return undefined;
+
+  if (auctionManagerInstance) {
+    // instead we apply defective state to auction managers
+    if (
+      desiredState === AuctionViewState.Defective &&
+      auctionManagerInstance.info.state.status !==
+        AuctionManagerStatus.Initialized
+    )
+      return undefined;
+    // Generally the only way an initialized auction manager can get through is if you are asking for defective ones.
+    else if (
+      desiredState !== AuctionViewState.Defective &&
+      auctionManagerInstance.info.state.status ===
+        AuctionManagerStatus.Initialized
+    )
+      return undefined;
+
+    const vault = vaults[auctionManagerInstance.info.vault];
+    const auctionManagerKey = auctionManagerInstance.pubkey;
+
+    let safetyDepositConfigs: ParsedAccount<SafetyDepositConfig>[] =
+      buildListWhileNonZero(
+        safetyDepositConfigsByAuctionManagerAndIndex,
+        auctionManagerKey,
+      );
+
+    let bidRedemptions: ParsedAccount<BidRedemptionTicketV2>[] =
+      buildListWhileNonZero(
+        bidRedemptionV2sByAuctionManagerAndWinningIndex,
+        auctionManagerKey,
+      );
+    const auctionManager = new AuctionManager({
+      instance: auctionManagerInstance,
+      auction,
+      vault,
+      safetyDepositConfigs,
+      bidRedemptions,
+    });
+
+    const boxesExpected = auctionManager.safetyDepositBoxesExpected.toNumber();
+
+    let bidRedemption: ParsedAccount<BidRedemptionTicket> | undefined =
+      cachedRedemptionKeysByWallet[auction.pubkey]?.info
+        ? (cachedRedemptionKeysByWallet[
+            auction.pubkey
+          ] as ParsedAccount<BidRedemptionTicket>)
+        : undefined;
+
+    const bidderMetadata =
+      bidderMetadataByAuctionAndBidder[auction.pubkey + '-' + walletPubkey];
+    const bidderPot =
+      bidderPotsByAuctionAndBidder[auction.pubkey + '-' + walletPubkey];
+
+    if (existingAuctionView && existingAuctionView.totallyComplete) {
+      // If totally complete, we know we arent updating anythign else, let's speed things up
+      // and only update the two things that could possibly change
+      existingAuctionView.myBidderPot = bidderPot;
+      existingAuctionView.myBidderMetadata = bidderMetadata;
+      existingAuctionView.myBidRedemption = bidRedemption;
+      for (let i = 0; i < existingAuctionView.items.length; i++) {
+        const winningSet = existingAuctionView.items[i];
+        for (let j = 0; j < winningSet.length; j++) {
+          const curr = winningSet[j];
+          if (!curr.metadata) {
+            let foundMetadata =
+              metadataByMint[curr.safetyDeposit.info.tokenMint];
+            if (!foundMetadata) {
+              // Means is a limited edition, so the tokenMint is the printingMint
+              let masterEdition =
+                masterEditionsByPrintingMint[curr.safetyDeposit.info.tokenMint];
+              if (masterEdition) {
+                foundMetadata = metadataByMasterEdition[masterEdition.pubkey];
+              }
+            }
+            curr.metadata = foundMetadata;
+          }
+
+          if (
+            curr.metadata &&
+            !curr.masterEdition &&
+            curr.metadata.info.masterEdition
+          ) {
+            let foundMaster = masterEditions[curr.metadata.info.masterEdition];
+
+            curr.masterEdition = foundMaster;
+          }
+        }
+      }
+
+      return existingAuctionView;
+    }
+
+    const vaultKey = auctionManager.vault;
+    let boxes: ParsedAccount<SafetyDepositBox>[] = buildListWhileNonZero(
+      safetyDepositBoxesByVaultAndIndex,
+      vaultKey,
+    );
+    if (boxes.length > 0) {
+      let participationMetadata: ParsedAccount<Metadata> | undefined =
+        undefined;
+      let participationBox: ParsedAccount<SafetyDepositBox> | undefined =
+        undefined;
+      let participationMaster:
+        | ParsedAccount<MasterEditionV1 | MasterEditionV2>
+        | undefined = undefined;
+      if (
+        auctionManager.participationConfig !== null &&
+        auctionManager.participationConfig !== undefined
+      ) {
+        participationBox =
+          boxes[auctionManager.participationConfig?.safetyDepositBoxIndex];
+        // Cover case of V1 master edition (where we're using one time auth mint in storage)
+        // and case of v2 master edition where the edition itself is stored
+        participationMetadata =
+          metadataByMasterEdition[
+            masterEditionsByOneTimeAuthMint[participationBox.info.tokenMint]
+              ?.pubkey
+          ] || metadataByMint[participationBox.info.tokenMint];
+        if (participationMetadata) {
+          participationMaster =
+            masterEditionsByOneTimeAuthMint[participationBox.info.tokenMint] ||
+            (participationMetadata.info.masterEdition &&
+              masterEditions[participationMetadata.info.masterEdition]);
+        }
+      }
+
+      let view: Partial<AuctionView> = {
+        auction,
+        auctionManager,
+        state,
+        vault,
+        safetyDepositBoxes: boxes,
+        items: auctionManager.getItemsFromSafetyDepositBoxes(
+          metadataByMint,
+          masterEditionsByPrintingMint,
+          metadataByMasterEdition,
+          masterEditions,
+          boxes,
+        ),
+        participationItem:
+          participationMetadata && participationBox
+            ? {
+                metadata: participationMetadata,
+                safetyDeposit: participationBox,
+                masterEdition: participationMaster,
+                amount: new BN(1),
+                winningConfigType: WinningConfigType.Participation,
+              }
+            : undefined,
+        myBidderMetadata: bidderMetadata,
+        myBidderPot: bidderPot,
+        myBidRedemption: bidRedemption,
+      };
+
+      view.thumbnail =
+        ((view.items || [])[0] || [])[0] || view.participationItem;
+
+      view.totallyComplete = !!(
+        view.thumbnail &&
+        boxesExpected ===
+          (view.items || []).length +
+            (auctionManager.participationConfig === null ||
+            auctionManager.participationConfig === undefined
+              ? 0
+              : 1) &&
+        (auctionManager.participationConfig === null ||
+          auctionManager.participationConfig === undefined ||
+          (auctionManager.participationConfig !== null &&
+            view.participationItem)) &&
+        view.vault
+      );
+      if (
+        (!view.thumbnail || !view.thumbnail.metadata) &&
+        desiredState !== AuctionViewState.Defective
+      )
+        return undefined;
+
+      return view as AuctionView;
+    }
+  }
+
+  return undefined;
 }
