+++ js/packages/web/src/actions/decommAuctionManagerAndReturnPrizes.ts
@@ -5,10 +5,10 @@ import {
   sendTransactionsWithManualRetry,
   setAuctionAuthority,
   setVaultAuthority,
+  AuctionView,
   WalletSigner,
 } from '@oyster/common';
 import { WalletNotConnectedError } from '@solana/wallet-adapter-base';
-import { AuctionView } from '../hooks';
 import { AuctionManagerStatus } from '@oyster/common/dist/lib/models/metaplex/index';
 import { decommissionAuctionManager } from '@oyster/common/dist/lib/models/metaplex/decommissionAuctionManager';
 import { unwindVault } from './unwindVault';
