+++ js/packages/web/src/actions/sendRedeemBid.ts
@@ -22,6 +22,7 @@ import {
   BidderMetadata,
   getEditionMarkPda,
   decodeEditionMarker,
+  AuctionView,
   BidStateType,
   StringPublicKey,
   toPublicKey,
@@ -29,7 +30,6 @@ import {
 } from '@oyster/common';
 import { WalletNotConnectedError } from '@solana/wallet-adapter-base';
 import { AccountLayout, MintLayout, Token } from '@solana/spl-token';
-import { AuctionView } from '../hooks';
 import {
   AuctionManagerV1,
   ParticipationStateV1,
@@ -866,7 +866,7 @@ export async function setupRedeemParticipationInstructions(
       payingSolAccount,
       mint,
       me.info.supply.add(new BN(1)),
-      winnerIndex != null && winnerIndex != undefined
+      winnerIndex !== null && winnerIndex !== undefined
         ? new BN(winnerIndex)
         : null,
       myInstructions,
