+++ js/packages/web/src/components/Notifications/index.tsx
@@ -11,6 +11,7 @@ import {
   toPublicKey,
   useConnection,
   useUserAccounts,
+  AuctionViewState,
   VaultState,
   WalletSigner,
 } from '@oyster/common';
@@ -27,7 +28,7 @@ import { settle } from '../../actions/settle';
 import { startAuctionManually } from '../../actions/startAuctionManually';
 import { QUOTE_MINT } from '../../constants';
 import { useMeta } from '../../contexts';
-import { AuctionViewState, useAuctions } from '../../hooks';
+import { useAuctions } from '../../hooks';
 
 interface NotificationCard {
   id: string;
