+++ js/packages/web/src/hooks/useCreators.ts
@@ -1,7 +1,7 @@
 import { useMemo } from 'react';
 import { useMeta } from '../contexts';
 import { Artist } from '../types';
-import { AuctionView } from './useAuctions';
+import { AuctionView } from '@oyster/common';
 
 export const useCreators = (auction?: AuctionView) => {
   const { whitelistedCreatorsByCreator } = useMeta();
